update work in progress derivation pipeline with todo spec per function
